barra de espera de turno

+++ Assets/Resources/Scripts/DestroyDelay.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class DestroyDelay : MonoBehaviour
 {
-    [SerializeField] private int destroyTime = 1;
+    [SerializeField] private float destroyTime = 1f;
     private void Awake()
     {
         Invoke(nameof(Destroy), destroyTime);
